// finance_client/src/app/user/data/types/supplier.ts
export interface Supplier {
  deletedBy?: string;
  deletedFlag?:string;
  deletedTime?: Date;
  id?: 0;
  modifiedBy?: string;
  modifiedFlag?: string;
  modifiedTime?: Date;
  postedBy?: string;
  postedFlag?: string;
  postedTime?: Date;
  supplierAccount?: string;
  supplierAddress?: string;
  supplierBank?: string;
  supplierContact?: 0;
  supplierEmail?: string;
  supplierName?: string;
  supplierNumber?: string;
  supplierServices?: string;
  supplierTin?: string;
  supplierCountry?: string;
  supplierCurrency?: string;
  verifiedBy?: string;
  verifiedFlag?: string;
  verifiedTime?: Date;
}